refactor: rename rangeDatePicker to dateRangePicker
